chore(stockfish): remove legacy HTTP, fix binaries, update frontend socket
