+++ stockfishServer/bin/stockfish_11_mac
